Moved Entity to the Support namespace
